inline multiselect helpers
